[gandalff] DataFrame base test: old tests added back.

+++ core/gandalff/gdl_dataframe_base_bench_test.go
@@ -99,7 +99,7 @@ func Benchmark_Filter_Q1_1e5(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		df.Filter(
 			df.Series("id6").Map(func(v any) any {
-				return v.(int) > 500
+				return v.(int64) > 500
 			}, nil).(SeriesBool).Or(
 				df.Series("id1").Map(func(v any) any {
 					return v.(string) == "id024"
@@ -120,7 +120,7 @@ func Benchmark_Filter_Q1_1e6(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		df.Filter(
 			df.Series("id6").Map(func(v any) any {
-				return v.(int) > 500
+				return v.(int64) > 500
 			}, nil).(SeriesBool).Or(
 				df.Series("id1").Map(func(v any) any {
 					return v.(string) == "id024"
@@ -141,7 +141,7 @@ func Benchmark_Filter_Q1_1e7(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		df.Filter(
 			df.Series("id6").Map(func(v any) any {
-				return v.(int) > 500
+				return v.(int64) > 500
 			}, nil).(SeriesBool).Or(
 				df.Series("id1").Map(func(v any) any {
 					return v.(string) == "id024"
@@ -162,7 +162,7 @@ func Benchmark_Filter_Q2_1e5(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		df.Filter(
 			df.Series("id6").Map(func(v any) any {
-				return v.(int) > 500
+				return v.(int64) > 500
 			}, nil).(SeriesBool).And(
 				df.Series("v3").Map(func(v any) any {
 					return v.(float64) < 50
@@ -177,11 +177,11 @@ func Benchmark_Filter_Q2_1e5(b *testing.B) {
 				),
 			).(SeriesBool).And(
 				df.Series("v1").Map(func(v any) any {
-					return v.(int) == 5
+					return v.(int64) == 5
 				}, nil).(SeriesBool),
 			).(SeriesBool).And(
 				df.Series("v2").Map(func(v any) any {
-					return v.(int) == 1
+					return v.(int64) == 1
 				}, nil).(SeriesBool),
 			).(SeriesBool),
 		)
@@ -201,7 +201,7 @@ func Benchmark_Filter_Q2_1e6(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		df.Filter(
 			df.Series("id6").Map(func(v any) any {
-				return v.(int) > 500
+				return v.(int64) > 500
 			}, nil).(SeriesBool).And(
 				df.Series("v3").Map(func(v any) any {
 					return v.(float64) < 50
@@ -216,11 +216,11 @@ func Benchmark_Filter_Q2_1e6(b *testing.B) {
 				),
 			).(SeriesBool).And(
 				df.Series("v1").Map(func(v any) any {
-					return v.(int) == 5
+					return v.(int64) == 5
 				}, nil).(SeriesBool),
 			).(SeriesBool).And(
 				df.Series("v2").Map(func(v any) any {
-					return v.(int) == 1
+					return v.(int64) == 1
 				}, nil).(SeriesBool),
 			).(SeriesBool),
 		)
@@ -240,7 +240,7 @@ func Benchmark_Filter_Q2_1e7(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		df.Filter(
 			df.Series("id6").Map(func(v any) any {
-				return v.(int) > 500
+				return v.(int64) > 500
 			}, nil).(SeriesBool).And(
 				df.Series("v3").Map(func(v any) any {
 					return v.(float64) < 50
@@ -255,11 +255,11 @@ func Benchmark_Filter_Q2_1e7(b *testing.B) {
 				),
 			).(SeriesBool).And(
 				df.Series("v1").Map(func(v any) any {
-					return v.(int) == 5
+					return v.(int64) == 5
 				}, nil).(SeriesBool),
 			).(SeriesBool).And(
 				df.Series("v2").Map(func(v any) any {
-					return v.(int) == 1
+					return v.(int64) == 1
 				}, nil).(SeriesBool),
 			).(SeriesBool),
 		)
